+++ src/main/java/net/litetex/authback/mixin/common/YggdrasilMinecraftSessionServiceMixin.java
@@ -5,7 +5,6 @@ import java.util.Set;
 import java.util.UUID;
 
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -29,14 +28,14 @@ import com.mojang.util.UndashedUuid;
 
 import net.litetex.authback.common.AuthBackCommon;
 import net.litetex.authback.common.gameprofile.GameProfileCacheManager;
+import net.litetex.authback.shared.mixin.log.MixinLogger;
 
 
 @Mixin(value = YggdrasilMinecraftSessionService.class, remap = false)
 public abstract class YggdrasilMinecraftSessionServiceMixin
 {
 	@Unique
-	private static final Logger LOG =
-		LoggerFactory.getLogger("net.litetex.injected.ABYggdrasilMinecraftSessionService");
+	private static final Logger LOG = MixinLogger.common("YggdrasilMinecraftSessionServiceMixin");
 	
 	@Final
 	@Shadow