+++ src/main/java/net/litetex/authback/mixin/common/YggdrasilServicesKeyInfoMixin.java
@@ -5,7 +5,6 @@ import java.time.Instant;
 import java.util.Optional;
 
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -17,13 +16,14 @@ import com.mojang.authlib.yggdrasil.YggdrasilServicesKeyInfo;
 
 import net.litetex.authback.common.AuthBackCommon;
 import net.litetex.authback.common.GlobalPublicKeysCache;
+import net.litetex.authback.shared.mixin.log.MixinLogger;
 
 
 @Mixin(value = YggdrasilServicesKeyInfo.class, remap = false)
 public abstract class YggdrasilServicesKeyInfoMixin
 {
 	@Unique
-	private static final Logger LOG = LoggerFactory.getLogger(YggdrasilServicesKeyInfoMixin.class);
+	private static final Logger LOG = MixinLogger.common("YggdrasilServicesKeyInfoMixin");
 	
 	@Redirect(
 		method = "fetch",