+++ src/main/java/net/litetex/authback/mixin/server/DedicatedServerMixin.java
@@ -1,7 +1,6 @@
 package net.litetex.authback.mixin.server;
 
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -9,6 +8,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import net.litetex.authback.server.AuthBackServer;
+import net.litetex.authback.shared.mixin.log.MixinLogger;
 import net.minecraft.server.dedicated.DedicatedServer;
 
 
@@ -16,7 +16,7 @@ import net.minecraft.server.dedicated.DedicatedServer;
 public abstract class DedicatedServerMixin
 {
 	@Unique
-	private static final Logger LOG = LoggerFactory.getLogger(DedicatedServerMixin.class);
+	private static final Logger LOG = MixinLogger.server("DedicatedServerMixin");
 	
 	@Inject(
 		method = "convertOldUsers",